refactor: split os app agent bootstrap helpers
