+++ crates/temper-platform/src/os_apps/mod.rs
@@ -22,6 +22,7 @@ use temper_spec::csdl::{emit_csdl_xml, merge_csdl, parse_csdl};
 use crate::bootstrap;
 use crate::state::PlatformState;
 
+mod agent_bootstrap;
 pub mod git_sources;
 mod system_files;
 mod types;
@@ -1519,8 +1520,14 @@ async fn install_os_app_without_dependencies(
     let app_id = bootstrap_app_entity(state, &tenant_id, tenant, app_name).await;
 
     // ── Step 6: Bootstrap agents (returns name→uuid map). ──────────────
-    let (agents_bootstrapped, agent_uuid_map) =
-        bootstrap_agents(state, &tenant_id, tenant, &bundle.agents, app_id.as_deref()).await;
+    let (agents_bootstrapped, agent_uuid_map) = agent_bootstrap::bootstrap_agents(
+        state,
+        &tenant_id,
+        tenant,
+        &bundle.agents,
+        app_id.as_deref(),
+    )
+    .await;
 
     // ── Step 7: Bootstrap skills (agent-scoped + system). ──────────────
     let skills_bootstrapped =
@@ -1748,390 +1755,6 @@ async fn bootstrap_app_entity(
             None
         }
     }
-}
-
-/// Bootstrap agent definitions into the tenant by creating Soul entities.
-///
-/// For each agent definition in the app's `agents/` directory:
-/// 1. Check if the Soul entity type is registered (skip gracefully if not)
-/// 2. Check if a Soul with this name already exists (idempotent — name-based lookup)
-/// 3. Create a TemperFS File entity with the agent's concatenated content
-/// 4. Create a Soul entity pointing to that file (Temper auto-assigns prefixed UUID)
-/// 5. Set `source_app_id` on the Agent entity if App entity exists
-///
-/// Returns (names of bootstrapped agents, map of agent_name → soul entity UUID).
-async fn bootstrap_agents(
-    state: &PlatformState,
-    tenant_id: &TenantId,
-    tenant: &str,
-    agents: &[AgentDefinition],
-    app_id: Option<&str>,
-) -> (Vec<String>, BTreeMap<String, String>) {
-    let mut bootstrapped = Vec::new();
-    let mut agent_uuid_map: BTreeMap<String, String> = BTreeMap::new();
-
-    if agents.is_empty() {
-        return (bootstrapped, agent_uuid_map);
-    }
-
-    // Check if Soul entity type is registered.
-    let has_souls = {
-        let registry = state.registry.read().unwrap(); // ci-ok: infallible lock
-        registry.get_spec(tenant_id, "Soul").is_some()
-    };
-    if !has_souls {
-        tracing::info!(
-            tenant,
-            count = agents.len(),
-            "Skipping agent bootstrap — Soul entity type not registered (install paw-agent first)"
-        );
-        return (bootstrapped, agent_uuid_map);
-    }
-
-    // Check if File entity type is registered (for TemperFS).
-    let has_files = {
-        let registry = state.registry.read().unwrap(); // ci-ok: infallible lock
-        registry.get_spec(tenant_id, "File").is_some()
-    };
-    if !has_files {
-        tracing::info!(
-            tenant,
-            "Skipping agent bootstrap — File entity type not registered (install temper-fs first)"
-        );
-        return (bootstrapped, agent_uuid_map);
-    }
-
-    // Check if Agent entity type is registered (for source_app_id).
-    let has_agents = {
-        let registry = state.registry.read().unwrap(); // ci-ok: infallible lock
-        registry.get_spec(tenant_id, "Agent").is_some()
-    };
-
-    let agent_ctx = temper_server::request_context::AgentContext::system();
-
-    for agent in agents {
-        let stable_soul_id = bootstrapped_agent_soul_entity_id(&agent.name);
-
-        if let Ok(resp) = state
-            .server
-            .get_tenant_entity_state(tenant_id, "Soul", &stable_soul_id)
-            .await
-        {
-            if let Some(file_id) = resp
-                .state
-                .fields
-                .get("ContentFileId")
-                .or_else(|| resp.state.fields.get("content_file_id"))
-                .and_then(|v| v.as_str())
-            {
-                let desired_hash = content_sha256(agent.content.as_bytes());
-                match inspect_agent_soul_refresh(state, tenant_id, file_id, &desired_hash).await {
-                    Ok(AgentSoulRefreshDecision::AlreadyCurrent) => {
-                        tracing::debug!(
-                            tenant,
-                            agent = %agent.name,
-                            soul_id = %stable_soul_id,
-                            file_id = %file_id,
-                            "Stable agent soul already matches app bundle"
-                        );
-                    }
-                    Ok(AgentSoulRefreshDecision::PreserveCustomized) => {
-                        tracing::info!(
-                            tenant,
-                            agent = %agent.name,
-                            soul_id = %stable_soul_id,
-                            file_id = %file_id,
-                            "Preserving customized stable agent soul content"
-                        );
-                    }
-                    Ok(AgentSoulRefreshDecision::Upload) => {
-                        if let Err(error) = ensure_inline_file_uploaded(
-                            state,
-                            tenant_id,
-                            &agent_ctx,
-                            file_id,
-                            &format!("{}.soul.md", slugify_bootstrapped_agent_name(&agent.name)),
-                            agent.content.as_bytes(),
-                        )
-                        .await
-                        {
-                            tracing::warn!(
-                                tenant,
-                                agent = %agent.name,
-                                soul_id = %stable_soul_id,
-                                file_id = %file_id,
-                                error = %error,
-                                "Failed to refresh stable agent soul content"
-                            );
-                        }
-                    }
-                    Err(error) => {
-                        tracing::warn!(
-                            tenant,
-                            agent = %agent.name,
-                            soul_id = %stable_soul_id,
-                            file_id = %file_id,
-                            error = %error,
-                            "Failed to inspect stable agent soul content"
-                        );
-                    }
-                }
-            }
-
-            agent_uuid_map.insert(agent.name.clone(), stable_soul_id.clone());
-            bootstrapped.push(agent.name.clone());
-            if has_agents && let Some(app_id) = app_id {
-                set_agent_source_app(state, tenant_id, tenant, &agent_ctx, &agent.name, app_id)
-                    .await;
-            }
-            continue;
-        }
-
-        // Name-based lookup: find existing Soul with this name.
-        let existing_ids = state.server.list_entity_ids(tenant_id, "Soul");
-        let mut found_soul_id: Option<String> = None;
-        for id in &existing_ids {
-            if let Ok(resp) = state
-                .server
-                .get_tenant_entity_state(tenant_id, "Soul", id)
-                .await
-                && let Some(name) = state_field_str(&resp.state.fields, &["Name", "name"])
-                && name.eq_ignore_ascii_case(&agent.name)
-            {
-                // Update existing content file.
-                if let Some(file_id) = resp
-                    .state
-                    .fields
-                    .get("ContentFileId")
-                    .or_else(|| resp.state.fields.get("content_file_id"))
-                    .and_then(|v| v.as_str())
-                {
-                    let desired_hash = content_sha256(agent.content.as_bytes());
-                    match inspect_agent_soul_refresh(state, tenant_id, file_id, &desired_hash).await
-                    {
-                        Ok(AgentSoulRefreshDecision::AlreadyCurrent) => {
-                            tracing::debug!(
-                                tenant,
-                                agent = %agent.name,
-                                file_id = %file_id,
-                                "Existing agent soul content already matches app bundle"
-                            );
-                        }
-                        Ok(AgentSoulRefreshDecision::PreserveCustomized) => {
-                            tracing::info!(
-                                tenant,
-                                agent = %agent.name,
-                                file_id = %file_id,
-                                "Preserving customized existing agent soul content"
-                            );
-                        }
-                        Ok(AgentSoulRefreshDecision::Upload) => {
-                            if let Err(error) = ensure_inline_file_uploaded(
-                                state,
-                                tenant_id,
-                                &agent_ctx,
-                                file_id,
-                                &format!("{}.soul.md", agent.name.to_lowercase().replace(' ', "-")),
-                                agent.content.as_bytes(),
-                            )
-                            .await
-                            {
-                                tracing::warn!(
-                                    tenant,
-                                    agent = %agent.name,
-                                    file_id = %file_id,
-                                    error = %error,
-                                    "Failed to refresh existing agent soul content"
-                                );
-                            }
-                        }
-                        Err(error) => {
-                            tracing::warn!(
-                                tenant,
-                                agent = %agent.name,
-                                file_id = %file_id,
-                                error = %error,
-                                "Failed to inspect existing agent soul content"
-                            );
-                        }
-                    }
-                }
-                tracing::debug!(tenant, agent = %agent.name, soul_id = %id, "Soul already exists — updating");
-                found_soul_id = Some(id.clone());
-                break;
-            }
-        }
-
-        if let Some(soul_id) = found_soul_id {
-            agent_uuid_map.insert(agent.name.clone(), soul_id.clone());
-            bootstrapped.push(agent.name.clone());
-
-            // Set source_app_id on Agent entity if available.
-            if has_agents && let Some(app_id) = app_id {
-                set_agent_source_app(state, tenant_id, tenant, &agent_ctx, &agent.name, app_id)
-                    .await;
-            }
-            continue;
-        }
-
-        // Create new Soul — let Temper auto-assign a prefixed UUID.
-        // First, create the content file with a temporary ID (Temper will assign UUID).
-        let file_name = format!("{}.soul.md", slugify_bootstrapped_agent_name(&agent.name));
-        let temp_file_id = format!(
-            "bootstrap-soul-file-{}",
-            slugify_bootstrapped_agent_name(&agent.name)
-        );
-        if let Err(error) = ensure_inline_file_uploaded(
-            state,
-            tenant_id,
-            &agent_ctx,
-            &temp_file_id,
-            &file_name,
-            agent.content.as_bytes(),
-        )
-        .await
-        {
-            tracing::warn!(
-                tenant,
-                agent = %agent.name,
-                error = %error,
-                "Failed to create TemperFS File for agent"
-            );
-            continue;
-        }
-
-        // Create Soul entity with a stable prefixed ID so restarts stay idempotent
-        // even when entity indexes are still warming.
-        let new_soul_id = stable_soul_id;
-        match state
-            .server
-            .get_or_create_tenant_entity(tenant_id, "Soul", &new_soul_id, serde_json::json!({}))
-            .await
-        {
-            Ok(_) => {
-                let soul_id = new_soul_id;
-                // Initialize the Soul with metadata.
-                if let Err(e) = state
-                    .server
-                    .dispatch(temper_server::state::DispatchCommand {
-                        tenant: tenant_id,
-                        entity_type: "Soul",
-                        entity_id: &soul_id,
-                        action: "Create",
-                        params: serde_json::json!({
-                            "name": agent.name,
-                            "description": agent.description,
-                            "content_file_id": temp_file_id,
-                        }),
-                        agent_ctx: &agent_ctx,
-                        await_integration: false,
-                    })
-                    .await
-                {
-                    tracing::warn!(
-                        tenant,
-                        agent = %agent.name,
-                        error = %e,
-                        "Failed to register Soul entity"
-                    );
-                    continue;
-                }
-                // Publish the soul.
-                let _ = state
-                    .server
-                    .dispatch(temper_server::state::DispatchCommand {
-                        tenant: tenant_id,
-                        entity_type: "Soul",
-                        entity_id: &soul_id,
-                        action: "Publish",
-                        params: serde_json::json!({}),
-                        agent_ctx: &agent_ctx,
-                        await_integration: false,
-                    })
-                    .await;
-
-                tracing::info!(tenant, agent = %agent.name, soul_id = %soul_id, "Agent soul bootstrapped");
-                agent_uuid_map.insert(agent.name.clone(), soul_id);
-                bootstrapped.push(agent.name.clone());
-
-                // Set source_app_id on Agent entity if available.
-                if has_agents && let Some(app_id) = app_id {
-                    set_agent_source_app(state, tenant_id, tenant, &agent_ctx, &agent.name, app_id)
-                        .await;
-                }
-            }
-            Err(e) => {
-                tracing::warn!(
-                    tenant,
-                    agent = %agent.name,
-                    error = %e,
-                    "Failed to create Soul entity"
-                );
-            }
-        }
-    }
-    (bootstrapped, agent_uuid_map)
-}
-
-/// Set `source_app_id` on an Agent entity by looking it up by name.
-async fn set_agent_source_app(
-    state: &PlatformState,
-    tenant_id: &TenantId,
-    tenant: &str,
-    agent_ctx: &temper_server::request_context::AgentContext,
-    agent_name: &str,
-    app_id: &str,
-) {
-    // Find Agent entity by name.
-    let agent_ids = state.server.list_entity_ids(tenant_id, "Agent");
-    for id in &agent_ids {
-        if let Ok(resp) = state
-            .server
-            .get_tenant_entity_state(tenant_id, "Agent", id)
-            .await
-            && let Some(name) = state_field_str(&resp.state.fields, &["Name", "name"])
-            && name.eq_ignore_ascii_case(agent_name)
-        {
-            // Check if source_app_id is already set.
-            let current_app_id = resp
-                .state
-                .fields
-                .get("SourceAppId")
-                .or_else(|| resp.state.fields.get("source_app_id"))
-                .and_then(|v| v.as_str())
-                .unwrap_or_default();
-            if current_app_id == app_id {
-                return; // Already set correctly.
-            }
-            // Update via Configure action.
-            if let Err(e) = state
-                .server
-                .dispatch(temper_server::state::DispatchCommand {
-                    tenant: tenant_id,
-                    entity_type: "Agent",
-                    entity_id: id,
-                    action: "Configure",
-                    params: serde_json::json!({
-                        "source_app_id": app_id,
-                    }),
-                    agent_ctx,
-                    await_integration: false,
-                })
-                .await
-            {
-                tracing::warn!(
-                    tenant,
-                    agent = %agent_name,
-                    error = %e,
-                    "Failed to set source_app_id on Agent entity"
-                );
-            } else {
-                tracing::debug!(tenant, agent = %agent_name, app_id = %app_id, "Set source_app_id on Agent");
-            }
-            return;
-        }
-    }
-    tracing::debug!(tenant, agent = %agent_name, "No Agent entity found to set source_app_id");
 }
 
 /// Bootstrap skills as TemperFS files at path-based scope locations (ADR-002).
@@ -2403,125 +2026,16 @@ async fn bootstrap_skills(
     bootstrapped
 }
 
-async fn ensure_inline_file_uploaded(
-    state: &PlatformState,
-    tenant_id: &TenantId,
-    agent_ctx: &temper_server::request_context::AgentContext,
-    file_id: &str,
-    file_name: &str,
-    content: &[u8],
-) -> Result<(), String> {
-    let response = state
-        .server
-        .get_or_create_tenant_entity(tenant_id, "File", file_id, serde_json::json!({}))
-        .await
-        .map_err(|e| format!("failed to create File('{file_id}') actor: {e}"))?;
-
-    if response.state.status == "Created" {
-        state
-            .server
-            .dispatch(temper_server::state::DispatchCommand {
-                tenant: tenant_id,
-                entity_type: "File",
-                entity_id: file_id,
-                action: "Create",
-                params: serde_json::json!({
-                    "name": file_name,
-                    "path": "",
-                    "directory_id": "",
-                    "workspace_id": "",
-                    "mime_type": "text/markdown",
-                }),
-                agent_ctx,
-                await_integration: false,
-            })
-            .await
-            .map_err(|e| format!("failed to initialize File('{file_id}'): {e}"))?;
-    }
-
-    state
-        .server
-        .put_file_stream_content(tenant_id, file_id, content, "text/markdown", agent_ctx)
-        .await
-        .map_err(|e| format!("failed to upload File('{file_id}') content: {e}"))?;
-
-    Ok(())
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum AgentSoulRefreshDecision {
-    Upload,
-    AlreadyCurrent,
-    PreserveCustomized,
-}
-
-fn slugify_bootstrapped_agent_name(name: &str) -> String {
-    name.trim().to_lowercase().replace(' ', "-")
-}
-
-fn bootstrapped_agent_soul_entity_id(name: &str) -> String {
-    format!(
-        "sl-bootstrap-agent-soul-{}",
-        slugify_bootstrapped_agent_name(name)
-    )
-}
-
-async fn inspect_agent_soul_refresh(
-    state: &PlatformState,
-    tenant_id: &TenantId,
-    file_id: &str,
-    desired_hash: &str,
-) -> Result<AgentSoulRefreshDecision, String> {
-    let response = state
-        .server
-        .get_tenant_entity_state(tenant_id, "File", file_id)
-        .await
-        .map_err(|e| format!("failed to inspect File('{file_id}'): {e}"))?;
-
-    let has_content = response
-        .state
-        .booleans
-        .get("has_content")
-        .copied()
-        .unwrap_or(false);
-    let current_hash = response
-        .state
-        .fields
-        .get("content_hash")
-        .and_then(|value| value.as_str())
-        .unwrap_or_default();
-
-    Ok(decide_agent_soul_refresh(
-        has_content,
-        current_hash,
-        desired_hash,
-    ))
-}
-
-fn decide_agent_soul_refresh(
-    has_content: bool,
-    current_hash: &str,
-    desired_hash: &str,
-) -> AgentSoulRefreshDecision {
-    if !has_content || current_hash.is_empty() {
-        AgentSoulRefreshDecision::Upload
-    } else if current_hash == desired_hash {
-        AgentSoulRefreshDecision::AlreadyCurrent
-    } else {
-        AgentSoulRefreshDecision::PreserveCustomized
-    }
-}
-
-fn state_field_str<'a>(fields: &'a serde_json::Value, keys: &[&str]) -> Option<&'a str> {
-    keys.iter()
-        .find_map(|key| fields.get(*key).and_then(|value| value.as_str()))
-}
-
 pub(super) const APP_DOCS_WORKSPACE_ID: &str = "os-app-docs";
 const APP_DOCS_WORKSPACE_NAME: &str = "apps";
 pub(super) const APP_DOCS_ROOT_DIR_ID: &str = "os-app-docs-root";
 const APP_DOCS_ROOT_PATH: &str = "/apps";
 const APP_DOCS_QUOTA_BYTES: i64 = 1_099_511_627_776;
+
+fn state_field_str<'a>(fields: &'a serde_json::Value, keys: &[&str]) -> Option<&'a str> {
+    keys.iter()
+        .find_map(|key| fields.get(*key).and_then(|value| value.as_str()))
+}
 
 /// Bootstrap app-local ADRs into TemperFS under `/apps/{app-name}/adrs/`.
 async fn bootstrap_adrs(
